% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{c_matrix}
\alias{c_matrix}
\title{Confusion Matrix}
\usage{
c_matrix(y_pred, y_train, cutoff_value = 0.5)
}
\arguments{
\item{y_pred}{(this is the predicted value of target variable)}

\item{y_train}{(this is the actual target variable)}

\item{cutoff_value}{this is the cut off value. The default is 0.5}
}
\value{
Confusion matrix as table
}
\description{
Confusion Matrix
}
\author{
Shangjia
}
