% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Title}
\usage{
confusion_matrix(y_pred, y_train, cutoff_value = 0.5)
}
\arguments{
\item{y_pred}{(this is the predicted value of target variable)}

\item{y_train}{(this is the actual target variable)}

\item{cutoff_value}{this is the cut off value. The default is 0.5}
}
\value{
Confusion matrix, prevalence, accuracy, sensitivity, specificity, False Discovery Rate, Diagnostic Odds Ratio
}
\description{
Title
}
