% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{ConI}
\alias{ConI}
\title{Confidence Interval}
\usage{
ConI(alpha, x_train, y_train, B = 20)
}
\arguments{
\item{alpha}{\code{numeric} is the significance level to obtain the confidence intervals for beta}

\item{x_train}{\code{dataframe} or matrix (gets cast to matrix) that is our set of features.}

\item{y_train}{\code{dataframe} value of the target. Gets cast to matrix}

\item{B}{\code{numeric} of bootstraps which by default will be 20}
}
\description{
The is the function for confident interval by using Bootstrap
}
\author{
Shangjia Li
}
