% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{is_close}
\alias{is_close}
\title{I was trying something here, but I removed it because R wasn't happy with me}
\usage{
is_close(b1, b2)
}
\description{
I was trying something here, but I removed it because R wasn't happy with me
}
