% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{logistic_reg_predict_dataset}
\alias{logistic_reg_predict_dataset}
\title{Predicts Dataset}
\usage{
logistic_reg_predict_dataset(x_train, beta_cur)
}
\arguments{
\item{x_train}{\code{dataframe} or matrix (gets cast to matrix) that is our set of features.}

\item{beta_cur}{Numeric vector of length(number of features) that is found by logistic_regression_trainer}
}
\value{
Returns a column of predictions (the p_i). Every element is on interval (0,1). Use make_ones_and_zeros to get y_hats
}
\description{
Given a Beta, makes predictions on training data x_train
}
\author{
Noah Heckenlively
}
