% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{logistic_regression}
\alias{logistic_regression}
\title{This WILL OVERFIT to data, so it is prefered to get beta first and use your own.}
\usage{
logistic_regression(x_train, y_train, num_epochs = 20, lr = 1, cutoff = 0.5)
}
\arguments{
\item{x_train}{\code{dataframe} or matrix (gets cast to matrix) that is our set of features.}

\item{y_train}{\code{dataframe} value of the target. Gets cast to matrix}

\item{num_epochs}{\code{int} number of epochs to train for. Defaults to 20}

\item{lr}{\code{double} learning rate of logistic regression. If you are not converging, try lowering it.}

\item{cutoff}{A \code{double} on interval (0,1) that says to predict 1 if value >=cutoff. 0 otherwise}
}
\value{
Returns a column of predictions (1's and 0's) with cutoff of .5 by default
\describe{

}
}
\description{
Runs logistic regression on dataset.
}
\author{
Noah Heckenlively
}
