% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{logistic_regression_trainer}
\alias{logistic_regression_trainer}
\title{Trains Logistic Regression}
\usage{
logistic_regression_trainer(x_train, y_train, num_epochs = 20, lr = 1)
}
\arguments{
\item{x_train}{\code{dataframe} or matrix (gets cast to matrix) that is our set of features.}

\item{y_train}{\code{dataframe} value of the target. Gets cast to matrix}

\item{num_epochs}{\code{int} number of epochs to train for. Defaults to 20}

\item{lr}{\code{double} learning rate of logistic regression. If you are not converging, try lowering it.}
}
\value{
Returns Beta, a vector of length num_features that gets dotted with each row in sigmoid function during prediction.
}
\description{
Runs gradient descent to find beta
}
\author{
Noah Heckenlively
}
