% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{loss}
\alias{loss}
\title{Loss Function}
\usage{
loss(y_pred, y_train)
}
\arguments{
\item{y_pred}{column of p_i}

\item{y_train}{column of 1's and 0's}
}
\value{
Loss, a numeric type
}
\description{
Our loss function for the logistic regression
}
