% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{make_beta_init}
\alias{make_beta_init}
\title{Beta Init}
\usage{
make_beta_init(data, targs)
}
\arguments{
\item{data}{\code{matrix} of training data}

\item{targs}{\code{matrix} values of the target}
}
\value{
Returns Beta_init
}
\description{
Gets beta_init
}
