% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{Make_table}
\alias{Make_table}
\title{Accuracy Plots over Cutoff Values}
\usage{
Make_table(y_pred, y_train)
}
\arguments{
\item{y_pred}{\code{dataframe} or matrix (gets cast to matrix) that is our set of predictions.}

\item{y_train}{\code{dataframe} value of the target. Gets cast to matrix}
}
\description{
This function will provide you a plot of accuracy over a grid of cut-off values for prediction going from 0.1 to 0.9 with steps of 0.1.Partial Function is build up based on previous confusion_matrix function.
}
\author{
Shangjia Li
}
