% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{predict_row}
\alias{predict_row}
\title{Predict Row}
\usage{
predict_row(beta_cur, row)
}
\arguments{
\item{beta_cur}{Numeric vector of length(number of features) that used to predict target from data}

\item{row}{A row from out dataset that we are trying to predict on.}
}
\value{
P_i
}
\description{
Predicts target from the row. beta_cur and row must be same length because of dot product
}
\author{
Noah Heckenlively
}
