% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogiReg.R
\name{sigmoid}
\alias{sigmoid}
\title{Yeah, this feels a bit redundant
Why not just call this instead of predict_row?
    A: Because I could change predictor function if I really wanted by adding an optional
       variable to predict_row and letting user specify function}
\usage{
sigmoid(beta_cur, x_vec)
}
\arguments{
\item{beta_cur}{Numeric vector of length(number of features) that used to predict target from data}

\item{x_vec}{A row from out dataset that we are trying to predict on.}
}
\value{
P_i
}
\description{
Sigmoid function... what logistict regression uses to predict
}
\author{
Noah Heckenlively
}
